requirements.txt
pymongo==3.11.3
gTTS==2.2.2
discord.py==1.6.0
wordcloud==1.8.1
